fixed surface platform extensions,vulkan object cleanup order, warnings

+++ Sandbox/SandboxLayer.cpp
@@ -20,7 +20,7 @@ void SandboxLayer::OnUpdate(float dt)
 
     Renderer::DrawQuad(FRect{300.0f, 300.0f, 100.0f, 100.0f}, Color4{1.0f, 0.0f, 0.0f, 1.0f});
 
-    size_t length = snprintf(NULL, 0, "%u, %u", x, y);
+    size_t length = (size_t) snprintf(NULL, 0, "%u, %u", x, y);
     text[length] = '\0';
     snprintf(text, length + 1, "%d, %d", x, y);
